# openapi.yaml
openapi: 3.0.3
info:
  title: Bookings Engine API
  description: |
    Generic multi-tenant booking API for service-based businesses.
    Supports gyms, restaurants, salons, nail bars, and spas through
    configurable business rules rather than business-specific code paths.
  version: 1.0.0
  contact:
    name: Veltrion
servers:
  - url: https://api.bookings.example.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /resources:
    get:
      tags: [Resources]
      summary: List resources
      operationId: listResources
      parameters:
        - name: resource_type_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_active
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated list of resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Resource"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Resources]
      summary: Create a resource
      operationId: createResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceCreate"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Resources]
      summary: Get a resource
      operationId: getResource
      responses:
        "200":
          description: Resource details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    patch:
      tags: [Resources]
      summary: Update a resource
      operationId: updateResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    delete:
      tags: [Resources]
      summary: Deactivate a resource
      operationId: deleteResource
      responses:
        "204":
          description: Deactivated

  # ─── Resource Schedules ────────────────────────────────────────────────────────
  /resources/{id}/schedules:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Schedules]
      summary: Get weekly schedule for a resource
      operationId: getResourceSchedules
      responses:
        "200":
          description: Weekly schedule entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResourceSchedule"
    put:
      tags: [Schedules]
      summary: Replace weekly schedule
      operationId: setResourceSchedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedules]
              properties:
                schedules:
                  type: array
                  items:
                    $ref: "#/components/schemas/ResourceScheduleInput"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResourceSchedule"

  # ─── Schedule Overrides ────────────────────────────────────────────────────────
  /resources/{id}/overrides:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Schedules]
      summary: List schedule overrides
      operationId: listScheduleOverrides
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of overrides
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScheduleOverride"
    post:
      tags: [Schedules]
      summary: Create a schedule override
      operationId: createScheduleOverride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleOverrideCreate"
      responses:
        "201":
          description: Created override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleOverride"
  /resources/{id}/overrides/{overrideId}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
      - name: overrideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Schedules]
      summary: Delete a schedule override
      operationId: deleteScheduleOverride
      responses:
        "204":
          description: Deleted

  # ─── Services ──────────────────────────────────────────────────────────────────
  /services:
    get:
      tags: [Services]
      summary: List services
      operationId: listServices
      parameters:
        - name: resource_type_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_active
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated list of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Services]
      summary: Create a service
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    parameters:
      - $ref: "#/components/parameters/ServiceId"
    get:
      tags: [Services]
      summary: Get a service
      operationId: getService
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      tags: [Services]
      summary: Update a service
      operationId: updateService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      tags: [Services]
      summary: Deactivate a service
      operationId: deleteService
      responses:
        "204":
          description: Deactivated

  # ─── Resource-Service Links ────────────────────────────────────────────────────
  /resources/{id}/services:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Services]
      summary: List services a resource can perform
      operationId: getResourceServices
      responses:
        "200":
          description: Linked services
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
    put:
      tags: [Services]
      summary: Set services for a resource
      operationId: setResourceServices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_ids]
              properties:
                service_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Updated links
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"

  # ─── Availability ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ServiceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
  schemas:
...
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
...
    Resource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        resource_type_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ResourceCreate:
      type: object
      required: [resource_type_id, name]
      properties:
        resource_type_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        metadata:
          type: object
    ResourceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object

    # ─── Schedules ─────────────────────────────────────────────────────────────
    ResourceSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: "0 = Monday, 6 = Sunday"
        start_time:
          type: string
          description: "HH:mm format"
        end_time:
          type: string
          description: "HH:mm format"
        is_active:
          type: boolean
    ResourceScheduleInput:
      type: object
      required: [day_of_week, start_time, end_time]
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        start_time:
          type: string
        end_time:
          type: string
    ScheduleOverride:
      type: object
      properties:
        id:
          type: string
          format: uuid
        override_date:
          type: string
          format: date
        is_available:
          type: boolean
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        reason:
          type: string
    ScheduleOverrideCreate:
      type: object
      required: [override_date, is_available]
      properties:
        override_date:
          type: string
          format: date
        is_available:
          type: boolean
        start_time:
          type: string
        end_time:
          type: string
        reason:
          type: string

    # ─── Services ──────────────────────────────────────────────────────────────
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        duration_minutes:
          type: integer
        buffer_minutes:
          type: integer
        capacity:
          type: integer
        resource_type_id:
          type: string
          format: uuid
        price_cents:
          type: integer
        currency:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServiceCreate:
      type: object
      required: [name, duration_minutes, resource_type_id]
      properties:
        name:
          type: string
        description:
          type: string
        duration_minutes:
          type: integer
          minimum: 5
        buffer_minutes:
          type: integer
          minimum: 0
          default: 0
        capacity:
          type: integer
          minimum: 1
          default: 1
        resource_type_id:
          type: string
          format: uuid
        price_cents:
          type: integer
        currency:
          type: string
        metadata:
          type: object
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        duration_minutes:
          type: integer
          minimum: 5
        buffer_minutes:
          type: integer
          minimum: 0
        capacity:
          type: integer
          minimum: 1
        price_cents:
          type: integer
        currency:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object

    # ─── Availability ──────────────────────────────────────────────────────────
...